+++ app/models/slice/subject.rb
@@ -116,18 +116,29 @@ class Subject < SliceRecord
   end
 
   def generate_overview_report_pdf!(data)
-    jobname = "overview_report_#{id}"
-    output_folder = File.join("tmp", "files", "tex")
-    FileUtils.mkdir_p output_folder
-    file_tex = File.join("tmp", "files", "tex", "#{jobname}.tex")
+    jobname = "overview_report"
+    temp_dir = Dir.mktmpdir
+    temp_tex = File.join(temp_dir, "#{jobname}.tex")
+
+    write_tex_file(temp_tex, data)
+    self.class.compile(jobname, temp_dir, temp_tex)
+    temp_pdf = File.join(temp_dir, "#{jobname}.pdf")
+
+    @user.update overview_report_pdf: File.open(temp_pdf, "r"), overview_report_pdf_file_size: File.size(temp_pdf)
+  ensure
+    # Remove the directory.
+    FileUtils.remove_entry temp_dir
+  end
+
+  def write_tex_file(temp_tex, data)
+    # Needed by binding
     @subject = self
     @data = data
-    File.open(file_tex, "w") do |file|
+    File.open(temp_tex, "w") do |file|
       file.syswrite(ERB.new(latex_partial("header")).result(binding))
       file.syswrite(ERB.new(latex_partial("overview_report")).result(binding))
       file.syswrite(ERB.new(latex_partial("footer")).result(binding))
     end
-    Subject.generate_pdf(jobname, output_folder, file_tex)
   end
 
   private